Redmine #43: Fix zip file unpacking
We stopped automatic zip file unpacking as part of disabling MIME
detection, because it was broken.

This patch re-enables it, but *also* stores the original zip file.
So, when you upload a new zip file, you get one file stored called
"123.zip_dir" which is unpacked, and that's the file you're offered
by default, but we also upload "123.zip" which you can get at in
"Older files".