Switch to using password_hash for passwords
The previous code, which used MySQL's built-in AES_ENCRYPT, had a lot
of cryptographic problems. AES_ENCRYPT uses ECB, its key derivation
function is pretty lulzy, and it's completely unsalted.

The security requirements for Mystery Hunt aren't that substantial
(the likelihood of someone extracting and cracking the password
database seems low), but if someone re-uses a password on Puzzletron,
I'd like to store it in a safe way for the sake of the other sites
where they are using it.

This doesn't attempt to upgrade existing passwords, unless they're
changed. Everyone besides Death and Mayhem will probably never touch
their Puzzletron instance again, while we don't have any users yet.