Remove the "minimal email" option from puzzletron
As it turns out, "minimal email" actually means "no email", and it's
almost certainly never the case that you want that. Even if you think
you want that, I, as someone trying to run a hunt, don't want you to
be able to opt out of emails I want to send you.