Add a Slack notification when a puzzle is successfully testsolved
This has some embedded heuristics - specifically, that only one group
is likely to be solving a puzzle at a time, and so when any current
solver submits a correct answer, that means that all current solvers
collectively solved it.

It's probably good enough for now, and it means we don't have to worry
about de-bouncing multiple correct answer submissions, etc.